requirements.txt
mariadb>=1.1.10
PyYAML>=6.0.1
fastapi>=0.110.0
uvicorn>=0.27.1
requests>=2.31.0